process: make descriptor leak work-around conditional
This patch adds the __spawn_flags variable and __spawn_leak_workaround
flag to allow the software to control the leak work-around.
Previous behaviour was to always enable the work-around unless
the DPMI host is cwsdpmi.

Without this patch it is not possible to spawn a prot-mode TSR
program like 32rtm. djgpp treats it as a leak and wipes out of
memory. With this patch things work properly if the DPMI server
is smart enough to direct the control to prev client after 32rtm
TSRed. The problem is that 32rtm just jumps to the realmode exit
addr, so the DPMI server doesn't see the exit and may get confused
unless the special logic is implemented for that case (i.e. not all
DPMI servers treat that correctly even after that patch).