fix multiple definition errors with gcc 10
GCC 10 now defaults to -fno-common, which causes multiple definition
errors during linking.  In libc, __stdio_cleanup_hook is defined twice,
and in fsdb, several variables are defined in a header, which is
included in multiple source files.

To solve this, __stdio_cleanup_hook is declared extern and fsdb can be
compiled with -fcommon.