process: add _dos_exec5()
Same as _dos_exec() but with "flags" as 5th arg.
The flags are passed to AL when calling int21/4Bh.
In addition to an execution mode in AL, freedos defines the loadhigh
flag 0x80.

This allows to implement loadhigh properly.
Without this patch it is possible to implement loadhigh by hands
(by linking UMBs in a changing alloc strat), but in that case UMBs
will remain linked in.