dxe3gen: add support for using an exports file.
New command line option: --exports file

The existing mechanism for exported symbols was through symbol prefixes,
i.e. the -E and -X switches, but can be really hard to use especially if
there are many symbols which don't necessarily share a common prefix, or
if some of the symbols to exclude share the common prefix with symbols
to export.  After this, the -E and -X switches are still supported, but
those two are mutually exclusive with --exports.

The exports file is a plain text file with one symbol name at each line,
like an exports file that can be used in macOS projects.  Any leading or
trailing space/tab is ignored.  A line starting with a '#' character is
a comment line and will be ignored. Any missing exports are ignored and
reported to the user in verbose mode, unless --nomissing option is used.