get rid of dj64thnk.h
This header was shared between djdev64 and dj64's user-exported
code, which is not good.

To remove it, the dynamic registration of DJ64_DISPATCH_FN is used.
That way it is done behind djdev64's back, so it doesn't need that
define.
Unfortunately this makes the init sequence thread-unsafe, so protect
with the mutex.