dpmi: stop altering SS on __dpmi_int()
This seems to be unneeded, and seems to be the only place where
SS is altered w/o also switching ESP. That leads to a crash if the
code is executing on DPMI lpms, forcing the client to always switch
from lpms to some "emergency stack" first.
I am not sure why do we need that.