Fix typo
By accident, I've found a typo in the include/stdint.h file -
"FREESTANDIGN" instead of "FREESTANDING".

It doesn't do any harm. Yet. But, just in case, below is a patch for
the file. Since I wasn't able to find the problem (or the file) in the
source distribution, I'm patching the actual file instead.

Regards,
Bogdan Drozdowski